Rework FFmpeg options
- addresses an issue with usage of the FFmpeg option selection discussed on the Kodi forum https://forum.kodi.tv/showthread.php?tid=383046.
- added option to modify crf value

Note that choosing vaapi can be done via the browser but  could allow for setting v4l options in the future.
