+++ packages/addons/service/nextpvr/source/resources/Language/English/strings.po
@@ -38,11 +38,11 @@ msgid "Rescan m3u file(s) and update URLs"
 msgstr ""
 
 msgctxt "#30009"
-msgid "Apply transcoding selection"
+msgid "Transcode options"
 msgstr ""
 
 msgctxt "#30010"
-msgid "Change HLS transcoding mode, may not work on all platforms"
+msgid "Change FFmpeg video transcoding options for web streaming"
 msgstr ""
 
 msgctxt "#30011"
@@ -146,11 +146,19 @@ msgid "Delay service launch on boot to specified uptime (seconds)"
 msgstr ""
 
 msgctxt "#30051"
-msgid "Custom video preset"
+msgid "Choose video preset"
 msgstr ""
 
 msgctxt "#30052"
-msgid "Manage FFmpeg video transcoding settings"
+msgid "Manage FFmpeg video transcoding option"
+msgstr ""
+
+msgctxt "#30053"
+msgid "Set crf value"
+msgstr ""
+
+msgctxt "#30054"
+msgid "Constant Rate Factor (crf) quality. High to low with 18-28 recommended"
 msgstr ""
 
 msgctxt "#31000"
