+++ packages/addons/service/nextpvr/source/resources/settings.xml
@@ -20,7 +20,7 @@
       </group>
     </category>
     <category id="control" label="30004">
-      <group id="0">
+      <group id="1">
         <setting id="waitfor" type="integer" label="30049" help="30050">
           <level>1</level>
           <default>5</default>
@@ -54,30 +54,6 @@
             <close>false</close>
           </control>
         </setting>
-        <setting id="preset" type="string" label="30051" help="30052">
-            <level>2</level>
-            <default>copy</default>
-            <constraints>
-                <options>
-                    <option label="31000">copy</option>
-                    <option label="31001">default</option>
-                    <option label="31002">ultrafast</option>
-                    <option label="31003">superfast</option>
-                    <option label="31004">veryfast</option>
-                    <option label="31005">faster</option>
-                    <option label="31006">fast</option>
-                    <option label="31007">medium</option>
-                </options>
-            </constraints>
-            <control type="spinner" format="string"/>
-        </setting>
-        <setting id="transcode" type="action" label="30009" help="30010">
-          <level>2</level>
-          <data>RunScript(service.nextpvr, transcode)</data>
-          <control type="button" format="action">
-            <close>false</close>
-          </control>
-        </setting>
         <setting id="defaults" type="action" label="30044" help="30045">
           <level>3</level>
           <data>RunScript(service.nextpvr, defaults)</data>
@@ -93,6 +69,40 @@
           </control>
         </setting>
       </group>
+    </category>
+    <category id="ffmpeg" label="30009" help="30010">
+      <level type="ffmpeg">3</level>
+      <group id="3">
+        <setting id="preset" type="string" label="30051" help="30052">
+            <level>2</level>
+            <default>default</default>
+            <constraints>
+                <options>
+                    <option label="31000">copy</option>
+                    <option label="31001">default</option>
+                    <option label="31002">ultrafast</option>
+                    <option label="31003">superfast</option>
+                    <option label="31004">veryfast</option>
+                    <option label="31005">faster</option>
+                    <option label="31006">fast</option>
+                    <option label="31007">medium</option>
+                </options>
+            </constraints>
+            <control type="spinner" format="string"/>
+        </setting>
+        <setting id="crf" type="integer" label="30053" help="30054">
+          <level>2</level>
+          <default>26</default>
+          <constraints>
+            <minimum>0</minimum>
+            <step>1</step>
+            <maximum>51</maximum>
+          </constraints>
+          <control type="slider" format="integer">
+            <popup>false</popup>
+          </control>
+        </setting>
+        </group>
     </category>
   </section>
 </settings>
