+++ packages/addons/service/nextpvr/source/settings-default.xml
@@ -1,5 +1,5 @@
 <settings version="2">
     <setting id="waitfor" default="true">5</setting>
     <setting id="satiprtsp" default="true">554</setting>
-    <setting id="customvideo" default="true">-vcodec copy</setting>
+    <setting id="crf" default="true">26</setting>
 </settings>
